Struct turbo_malloc::TurboMalloc

pub struct TurboMalloc;

Turbo’s preferred global allocator. This is a new type instead of a type alias because you can’t use type aliases to instantiate unit types (E0423).

Implementations

impl TurboMalloc

pub fn memory_usage() -> usize

pub fn thread_stop()